check if directory exists before removing them.

+++ src/cmd/dev.ts
@@ -147,15 +147,19 @@ function _watch(){
 				
 			}else if(_event === 'unlink' || _event === 'unlinkDir'){
 				
-				if(util_instance.fs.is_directory(new_path_server)){
-					util_instance.fs.remove_directory(new_path_server);
-				}else{
-					util_instance.fs.remove_file(new_path_server);
+				if(util_instance.fs.exists(new_path_server)){
+					if(util_instance.fs.is_directory(new_path_server)){
+						util_instance.fs.remove_directory(new_path_server);
+					}else{
+						util_instance.fs.remove_file(new_path_server);
+					}
 				}
-				if(util_instance.fs.is_directory(new_path_client)){
-					util_instance.fs.remove_directory(new_path_client);
-				}else{
-					util_instance.fs.remove_file(new_path_client);
+				if(util_instance.fs.exists(new_path_client)){
+					if(util_instance.fs.is_directory(new_path_client)){
+						util_instance.fs.remove_directory(new_path_client);
+					}else{
+						util_instance.fs.remove_file(new_path_client);
+					}
 				}
 				
 			}else{
